fix: create wrapper folder and move HeroWrapper
