+++ client/src/pages/Home.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useQuery } from "@tanstack/react-query";
 import axios from "axios";
 import ProductCard from "../components/ProductCard";
-import HeroWrapper from "../components/HeroWrapper";
+import HeroWrapper from "../layout/HeroWrapper";
 
 const Home = () => {
   // Fetch products logic
